Update to latest version of HF Transformers (Fixes #4)
I hope this didn't break backwards compatibility in ways beyond what I
can test.

Anyways, lots of new stuff in HF Transformers, which make this quite a
lot harder than it was before. I'm not sure if I can continue supporting
this, if it this goes on. Perhaps they should think of a canonical way
of doing this...